Return clientKeys property/value into a string
return Object.keys(clientKeys)
  .map(key => `${key}=${clientKeys[key]}`)
  .join('&');

Object.keys returns an array of the given object's PROPERTY NAMES
in the same order as a normal loop.
Note:
  * when using an array, the array's index will be returned;
not the value.
  * when using numbers as the property name within an object {},
it will be rearranged from lowest to highest.

The map() method creates a new array with the results of calling
a provided function on every element in the calling array.
Note: must provide a function within the map method.

The join() method concatenate everything in the array into a single,
new string separates everything with the value inside the quotes.
Note: if no input <.join()> then by default,
it'll be commas with no spaces.